RavenDb-25281 - Code refactoring for clarity, separate files, better comments
