+++ src/Corax/Querying/Planning/PostingSourceKind.cs
@@ -1,0 +1,12 @@
+namespace Corax.Querying.Planning;
+
+public enum PostingSourceKind : byte
+{
+    /// <summary>The term does not exist in the index (or the field has no compact tree).
+    /// Dispatcher primitives no-op on Empty for Or-shaped ops, and clear the bitmap
+    /// for And-shaped ops.</summary>
+    Empty,
+    Single,
+    SmallPostingList,
+    PostingList,
+}
