+++ src/Corax/Querying/Planning/ScanCompareOp.cs
@@ -1,0 +1,12 @@
+namespace Corax.Querying.Planning;
+
+public enum ScanCompareOp : byte
+{
+    Equal,
+    NotEqual,
+    GreaterThan,
+    GreaterThanOrEqual,
+    LessThan,
+    LessThanOrEqual,
+    Between,
+}
