+++ src/Corax/Querying/Planning/ScanValueType.cs
@@ -1,0 +1,8 @@
+namespace Corax.Querying.Planning;
+
+public enum ScanValueType : byte
+{
+    Long,    // reader.CurrentLong vs ctx.LongParams[i]
+    Double,  // reader.CurrentDouble vs ctx.DoubleParams[i]
+    Slice,   // IL-emitted byte-sequence comparison
+}
