+++ src/Corax/Querying/Planning/SpatialParams.cs
@@ -1,0 +1,18 @@
+namespace Corax.Querying.Planning;
+
+/// <summary>Per-execution spatial query parameters. Resolved from ParameterBinding during
+/// PopulateClauseValues — scalar params looked up by name in the blittable.
+/// Shape construction (spatialField.ReadCircle / ReadShape) also runs at execution time
+/// because it needs the spatial field factory from builderParameters.</summary>
+public sealed class SpatialParams
+{
+    public double DistanceErrorPct = -1; // -1 = use default
+    public bool IsCircle;                // true = circle, false = WKT
+    // Circle parameters
+    public double CircleRadius;
+    public double CircleLatitude;
+    public double CircleLongitude;
+    // WKT parameter
+    public string Wkt;
+    public Utils.Spatial.SpatialUnits? Units;
+}
