+++ src/Corax/Querying/Planning/VectorSearchOp.cs
@@ -1,0 +1,11 @@
+namespace Corax.Querying.Planning;
+
+/// <summary>Vector nearest-neighbor search applied after the bitmap filter phase.
+/// The bitmap-producing match is passed as the filter source to VectorSearchMatch,
+/// restricting the search to the candidate set. MatchIndex is the slot in the
+/// resolved IQueryMatch[] for this vector search.</summary>
+public struct VectorSearchOp
+{
+    public ClauseInfo Clause;
+    public ClauseExecution Exec;
+}
